requirements.txt
imgui_bundle==1.6.3
moderngl==5.12.0
numpy==2.3.0
pyglm==2.8.2
trimesh==3.23.5
